Código finalizado

+++ controllers/product_controller.py
@@ -16,19 +16,26 @@ class ProductController(BaseController):
         self.app.route('/sahurhomepage', method='GET', callback=self.list_products)
         self.app.route('/products/add', method=['GET', 'POST'], callback=self.add_product)
         self.app.route('/products/edit/<product_id:int>', method=['GET', 'POST'], callback=self.edit_product)
-        self.app.route('/products/delete/<product_id:int>', method='POST', callback=self.delete_products)
+        self.app.route('/products/delete/<product_id:int>', method=['GET', 'POST'], callback=self.delete_products)
         self.app.route('/productsdetails/<product_id:int>', method='GET', callback=self.detail_products)
-
+        
+        
+       
         
 
-       
     
     def list_products(self):
         products = self.product_service.get_all()
         return self.render('home', products=products)
+    
+    
+
+
 
     def add_product(self):
         action_url = '/products/add'
+
+        self.check_auth()
         
         if request.method == 'GET':
             return self.render('product_form', product=None, action=action_url, error=None)
@@ -50,6 +57,8 @@ class ProductController(BaseController):
             
 
     def edit_product(self, product_id):
+        self.check_auth()
+        
         product = self.product_service.get_by_id(product_id)
 
         if not product:
@@ -62,8 +71,13 @@ class ProductController(BaseController):
             self.redirect('/sahurhomepage')
 
     def delete_products(self, product_id):
-        self.product_service.delete(product_id)
-        self.redirect('/sahurhomepage')
+        self.check_auth()
+        product = self.product_service.get_by_id(product_id)
+        if request.method == 'GET':
+            return self.render('product_delete_form', product=product, action=f'/products/delete/{product_id}')
+        else:
+            self.product_service.delete(product_id)
+            self.redirect('/sahurhomepage')
 
     
     def detail_products(self, product_id):
